refactor!: Modified configuration options
