+++ src/Admin/Views/_ViewImports.cshtml
@@ -2,6 +2,7 @@
 @using Microsoft.AspNetCore.Mvc.Localization
 @using Microsoft.Extensions.Options
 @using Nocturne.Auth.Admin.Configuration.Constants
+@using Nocturne.Auth.Admin.Configuration.Options
 @using Nocturne.Auth.Admin.Controllers.Models
 @using Nocturne.Auth.Core.Shared.Extensions
 @using Nocturne.Auth.Core.Web
