+++ src/Configuration/Options/ApplicationDataProtectionOptions.cs
@@ -1,7 +1,7 @@
 // Copyright (c) Leandro Silva Luz do Carmo
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-namespace Nocturne.Auth.Configuration.Models
+namespace Nocturne.Auth.Configuration.Options
 {
     public class ApplicationDataProtectionOptions
     {
