+++ src/Core/Shared/Extensions/StringExtensions.cs
@@ -35,7 +35,7 @@ namespace Nocturne.Auth.Core.Shared.Extensions
 
             return value.Length <= maxLength
                 ? value
-                : value.Substring(0, maxLength - truncateSufix.Length) + truncateSufix;
+                : value[.. (maxLength - truncateSufix.Length)] + truncateSufix;
         }
 
         public static string ToAlphaNumericOnly(this string value)
