+++ src/Server/Areas/Identity/Pages/Account/ResetPasswordConfirmation.cshtml.cs
@@ -9,7 +9,7 @@ namespace Nocturne.Auth.Server.Areas.Identity.Pages.Account
     [AllowAnonymous]
     public class ResetPasswordConfirmationModel : PageModel
     {
-        public void OnGet()
+        public static void OnGet()
         {
         }
     }
