+++ src/Server/Views/Shared/_Footer.cshtml
@@ -1,18 +1,18 @@
 @inject IViewLocalizer localizer;
-@inject IOptions<WebApplicationOptions> applicationOptions;
+@inject IOptions<ServerApplicationOptions> ApplicationOptions;
 
 @{
-    var currentApplicationOptions = applicationOptions.Value;
+    var applicationOptions = ApplicationOptions.Value;
 }
 
 <div class="layout-footer footer">
     <ul class="footer-list ml-auto">
-        @if (currentApplicationOptions.HasPrivacyPolicyUrl)
+        @if (applicationOptions.HasPrivacyPolicyUrl)
         {
             <li class="footer-list-item">
-                <a class="footer-link" href="@currentApplicationOptions.PrivacyPolicyUrl">@localizer["Privacy policy"]</a>
+                <a class="footer-link" href="@applicationOptions.PrivacyPolicyUrl">@localizer["Privacy policy"]</a>
             </li>
         }
-        <li class="footer-list-item">@currentApplicationOptions.CompanyName</li>
+        <li class="footer-list-item">@applicationOptions.CompanyName</li>
     </ul>
 </div>
