+++ src/Server/Views/_ViewImports.cshtml
@@ -2,4 +2,5 @@
 @using Microsoft.Extensions.Options
 @using Nocturne.Auth.Core.Web
 @using Nocturne.Auth.Server.Configuration.Constants
+@using Nocturne.Auth.Server.Configuration.Options
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
